Create pytest fixtures for ray_local_mode
Tensorflow policies can't be pickle and unpickled within
the same session. This means that if we were to run RaySampler
in local mode with tensorflow, the sampler will fail to
construct workers because the workers will try to create
policies whose underlying neural networks can't be copied
into the same session. This could potentially be fixed
if somehow upon unpickling the names of the variables
that make up the policy were to somehow change.